活動功能完成90%

+++ TGIS/Controllers/NormalOfferController.cs
@@ -11,53 +11,109 @@ namespace TGIS.Controllers
     public class NormalOfferController : Controller
     {
         TGISDBEntities db = new TGISDBEntities();
-        // GET: NormalOffer
+        //查看活動列表
         public ActionResult OfferList()
         {
+            //若有店家ID則顯示店家活動列表
+            Shop s = db.Shops.Find((string)Session["ShopID"]);
+            if (s != null)
+                return View("OfferList", "_LayoutShoperCenter", s.NormalOffers.ToList());
+            //否則顯示全部活動列表(管理員用)
             return View(db.NormalOffers.ToList());
         }
+
+        //店家新增優惠券
         public ActionResult OfferCreate()
         {
             return View();
         }
         [HttpPost]
-        public ActionResult OfferCreate(NormalOffer normalOffer, HttpPostedFileBase photo)
+        public ActionResult OfferCreate(NormalOffer normalOffer, HttpPostedFileBase[] photos)
         {
             if (ModelState.IsValid)
             {
+                //填入預設值
+                normalOffer.ID = UsefulTools.GetNextID(db.NormalOffers, 1);
                 normalOffer.ShopID = Session["ShopID"].ToString();
+                normalOffer.Clicks = 0;
+
                 db.NormalOffers.Add(normalOffer);
-                PhotoManager.Create(normalOffer.ID, new HttpPostedFileBase[] { photo });
                 db.SaveChanges();
+                //存入圖片
+                PhotoManager.Create(normalOffer.ID, photos);
                 return RedirectToAction("OfferList");
             }
             return View();
         }
-        public ActionResult OfferEdit(string id)
+
+        //店家編輯活動
+        public ActionResult OfferEdit(string normalOfferID)
         {
-            return View(db.NormalOffers.Find(id));
+            var offer = db.NormalOffers.Find(normalOfferID);
+            //檢查店家ID是否為活動所屬店家
+            string shopID = (string)Session["ShopID"];
+            if (shopID != offer.ShopID)
+                return HttpNotFound();
+
+            //將此活動的圖片數量傳入ViewBag
+            ViewBag.photoIDList = PhotoManager.GetPhotoIDList(normalOfferID);
+            TempData["Offer"] = offer;
+            return View(offer);
         }
         [HttpPost]
-        public ActionResult OfferEdit(NormalOffer normalOffer)
+        public ActionResult OfferEdit(NormalOffer normalOffer, HttpPostedFileBase[] photos)
         {
             if (ModelState.IsValid)
             {
+                //取出原始資料並灌入必要資料
+                NormalOffer offer = (NormalOffer)TempData["Offer"];
+                normalOffer.ID = offer.ID;
+                normalOffer.ShopID = offer.ShopID;
+                normalOffer.Clicks = offer.Clicks;
+
                 db.Entry(normalOffer).State = EntityState.Modified;
                 db.SaveChanges();
-                return View();
+                //存入圖片
+                PhotoManager.Create(normalOffer.ID, photos);
+                return RedirectToAction("OfferList");
             }
+            TempData.Keep("Offer");
             return View();
         }
-        public ActionResult OfferDel(string id)
+
+        //店家刪除活動
+        public ActionResult OfferDel(string normalOfferID)
         {
-            var offer = db.NormalOffers.Find(id);
+            var offer = db.NormalOffers.Find(normalOfferID);
+            //檢查店家ID是否為活動所屬店家
+            string shopID = (string)Session["ShopID"];
+            if (shopID != offer.ShopID)
+                return HttpNotFound();
+
+            //先刪除圖片
+            PhotoManager.Delete(normalOfferID);
             db.NormalOffers.Remove(offer);
             db.SaveChanges();
-            return View();
+            return RedirectToAction("OfferList");
         }
-        public ActionResult OfferDetail(string id)
+
+        //給玩家看的活動列表
+        [ChildActionOnly]
+        public ActionResult _OfferListForPlayer(string shopID)
         {
-            return View(db.NormalOffers.Find(id));
+            return PartialView(db.Shops.Find(shopID).NormalOffers.ToList());
         }
+
+        //給玩家看的活動詳細資料
+        public ActionResult OfferDetail(string normalOfferID)
+        {
+            var offer = db.NormalOffers.Find(normalOfferID);
+            //存入來源店家ID以便返回頁面
+            ViewBag.ShopID = offer.ShopID;
+            //將此活動的圖片數量傳入ViewBag
+            ViewBag.photoIDList = PhotoManager.GetPhotoIDList(normalOfferID);
+            return View(offer);
+        }
+
     }
 }
